gem-app-microservice/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Werkzeug==2.3.7
Jinja2==3.1.3
SQLAlchemy==2.0.23
itsdangerous>=2.0.0
click>=8.0.0
MarkupSafe>=2.0.0
blinker>=1.6.0
typing_extensions>=4.0.0
requests>=2.26.0
python-wordpress-xmlrpc>=2.3.0
PyPDF2>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
schedule>=1.0.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
boto3
awscli


